Accepts an alt fieldSpec when starting from Solr URL
This enables a feature to pass around the fieldSpec in the URL when
sharing splainer URLs. When you shared those URLs you lost the
information in the fieldspec, such as what your id field was and stuff
like function queries. Instead you just fell back onto whatever was in
Solr's fl which doesn't include info about the Splainer-specific roles
of each field